Added gamejams page, changed footer and misc project page layout
Added a Game Jams & Hackathons page that will just show an image for
each 'project' completed and link to external sites such as itch.io or
wherever the result from the gamejam/hackathon is stored.

Removed the footer containing contact information from the project
pages.

Changed the fluid container to just a static container containing both
the project media and the project description and links.