Project sections seperated, project page design changes, modal.
Seperated the projects by sections (Game dev, web and mobile, gamejams
and hackathons) that sit in wells on index.html.

Changed some design on the project page, project description is now
accompanied by a platform, made with, and timespan information in the
form of bullet points. The project links (itch.io, github and gamejolt)
now sit underneath the aforementioned information.

Images on the project page now pop up slightly enlarged (the effect
lessens the larger the screen is to begin with) within a modal.

Some misc css changes.